--- a/aiter/ops/triton/configs/gemm/MI350X-GEMM-A16W16-N=256-K=7168.json
+++ b/aiter/ops/triton/configs/gemm/MI350X-GEMM-A16W16-N=256-K=7168.json
@@ -9,6 +9,7 @@
     "waves_per_eu": 1,
     "matrix_instr_nonkdim": 16,
     "cache_modifier": ".cg",
+	"NUM_KSPLIT": 1,
     "kpack": 1
   },
   "M_LEQ_128": {
@@ -21,6 +22,7 @@
     "waves_per_eu": 1,
     "matrix_instr_nonkdim": 16,
     "cache_modifier": ".cg",
+	"NUM_KSPLIT": 1,
     "kpack": 1
   },
   "M_LEQ_256": {
@@ -33,6 +35,7 @@
     "waves_per_eu": 1,
     "matrix_instr_nonkdim": 32,
     "cache_modifier": null,
+	"NUM_KSPLIT": 1,
     "kpack": 1
   },
   "M_LEQ_512": {
@@ -45,6 +48,7 @@
     "waves_per_eu": 1,
     "matrix_instr_nonkdim": 32,
     "cache_modifier": null,
+	"NUM_KSPLIT": 1,
     "kpack": 1
   },
   "M_LEQ_2048": {
@@ -57,6 +61,7 @@
     "waves_per_eu": 1,
     "matrix_instr_nonkdim": 32,
     "cache_modifier": null,
+	"NUM_KSPLIT": 1,
     "kpack": 1
   },
   "M_GEQ_4096": {
@@ -69,6 +74,7 @@
     "waves_per_eu": 1,
     "matrix_instr_nonkdim": 32,
     "cache_modifier": null,
+	"NUM_KSPLIT": 1,
     "kpack": 1
   }
 }