## Motivation

Add splitK functionality to the GEMM A16W16 Triton Kernel.

## Technical Details

When NUMK_SPLIT = 1, performance is on par with the original kernel:

| | M | N | K | Original Time_(ms) | SplitK Time_(ms) |
| --- | --- | --- | --- | --- | --- |
| 0 | 1 | 1280 | 8192 | 0.098525 | 0.098311 |
| 1 | 32 | 1280 | 8192 | 0.102982 | 0.10255 |
| 2 | 64 | 1280 | 8192 | 0.106523 | 0.105869 |
| 3 | 128 | 1280 | 8192 | 0.117839 | 0.117624 |
| 4 | 192 | 1280 | 8192 | 0.119228 | 0.119421 |
| 5 | 256 | 1280 | 8192 | 0.119773 | 0.135714 |
| 6 | 320 | 1280 | 8192 | 0.117968 | 0.096636 |
| 7 | 512 | 1280 | 8192 | 0.099471 | 0.098889 |
| 8 | 1024 | 1280 | 8192 | 0.314774 | 0.375745 |
| 9 | 2048 | 1280 | 8192 | 0.319122 | 0.320532 |
| 10 | 4096 | 1280 | 8192 | 0.350127 | 0.349414 |
| 11 | 8192 | 1280 | 8192 | 0.436413 | 0.43571 |
| 12 | 16384 | 1280 | 8192 | 0.92344 | 0.928293 |


## Test Plan

<!-- Explain any relevant testing done to verify this PR. -->

## Test Result

Tests passed.

## Submission Checklist

- [ ] Look over the contributing guidelines at https://github.com/ROCm/ROCm/blob/develop/CONTRIBUTING.md#pull-requests.